Refactor inline query handler to separate controller file
